[C-5454] Add mobile web guest checkout flow
