+++ packages/web/src/components/premium-content-purchase-modal/PremiumContentPurchaseModal.tsx
@@ -29,16 +29,15 @@ import {
   purchaseContentActions,
   purchaseContentSelectors,
   PurchaseContentStage,
-  PurchaseContentPage,
+  PurchaseContentPage as PurchaseContentPageType,
   isContentPurchaseInProgress,
-  PurchaseableContentType
+  PurchaseableContentType,
+  accountSelectors
 } from '@audius/common/store'
-import { USDC } from '@audius/fixed-decimal'
 import {
   ModalContentPages,
   ModalHeader,
   ModalFooter,
-  Flex,
   IconCart,
   ModalTitle
 } from '@audius/harmony'
@@ -48,9 +47,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import { toFormikValidationSchema } from 'zod-formik-adapter'
 
 import { useHistoryContext } from 'app/HistoryProvider'
-import { GuestCheckout } from 'components/guest-checkout/GuestCheckout'
 import { ModalForm } from 'components/modal-form/ModalForm'
-import { LockedContentDetailsTile } from 'components/track/LockedContentDetailsTile'
 import { USDCManualTransfer } from 'components/usdc-manual-transfer/USDCManualTransfer'
 import { useIsMobile } from 'hooks/useIsMobile'
 import { useIsUSDCEnabled } from 'hooks/useIsUSDCEnabled'
@@ -60,45 +57,45 @@ import { pushUniqueRoute } from 'utils/route'
 import zIndex from 'utils/zIndex'
 
 import styles from './PremiumContentPurchaseModal.module.css'
-import { PurchaseContentFormFields } from './components/PurchaseContentFormFields'
 import { PurchaseContentFormFooter } from './components/PurchaseContentFormFooter'
 import { usePurchaseContentFormState } from './hooks/usePurchaseContentFormState'
-
+import {
+  GuestCheckoutFooter,
+  GuestCheckoutPage
+} from './pages/GuestCheckoutPage'
+import { PurchaseContentPage } from './pages/PurchaseContentPage'
 const { startRecoveryIfNecessary, cleanup: cleanupUSDCRecovery } =
   buyUSDCActions
 const { cleanup, setPurchasePage, eagerCreateUserBank } = purchaseContentActions
 const { getPurchaseContentFlowStage, getPurchaseContentError } =
   purchaseContentSelectors
+const { getHasAccount } = accountSelectors
 
 const messages = {
+  guestCheckout: 'Guest Checkout',
   completePurchase: 'Complete Purchase'
 }
 
-const pageToPageIndex = (page: PurchaseContentPage) => {
+const pageToPageIndex = (page: PurchaseContentPageType) => {
   switch (page) {
-    case PurchaseContentPage.GUEST_CHECKOUT:
+    case PurchaseContentPageType.GUEST_CHECKOUT:
       return 0
-    case PurchaseContentPage.PURCHASE:
+    case PurchaseContentPageType.PURCHASE:
       return 1
-    case PurchaseContentPage.TRANSFER:
+    case PurchaseContentPageType.TRANSFER:
       return 2
   }
 }
 
-// The bulk of the form rendering is in a nested component because we want access
-// to the FormikContext, which can only be used in a component which is a descendant
-// of the `<Formik />` component
-const RenderForm = ({
-  onClose,
-  metadata,
-  purchaseConditions,
-  contentId
-}: {
+type PremiumContentPurchaseFormProps = {
   onClose: () => void
   metadata: PurchaseableContentMetadata
   purchaseConditions: USDCPurchaseConditions
   contentId: ID
-}) => {
+}
+
+const PremiumContentPurchaseForm = (props: PremiumContentPurchaseFormProps) => {
+  const { onClose, metadata, purchaseConditions, contentId } = props
   const dispatch = useDispatch()
   const isMobile = useIsMobile()
   const { permalink } = metadata
@@ -109,13 +106,10 @@ const RenderForm = ({
     usePurchaseContentFormState({ price })
   const [, , { setValue: setPurchaseMethod }] = useField(PURCHASE_METHOD)
   const currentPageIndex = pageToPageIndex(page)
-  const isLinkDisabled =
-    stage === PurchaseContentStage.START ||
-    stage === PurchaseContentStage.PURCHASING ||
-    stage === PurchaseContentStage.CONFIRMING_PURCHASE
 
   const { submitForm, resetForm } = useFormikContext()
   const { history } = useHistoryContext()
+  const isSignedIn = useSelector(getHasAccount)
 
   // Reset form on track change
   useEffect(() => {
@@ -130,13 +124,14 @@ const RenderForm = ({
   }, [stage, permalink, dispatch, history])
 
   const handleUSDCManualTransferClose = useCallback(() => {
-    dispatch(setPurchasePage({ page: PurchaseContentPage.PURCHASE }))
+    dispatch(setPurchasePage({ page: PurchaseContentPageType.PURCHASE }))
   }, [dispatch])
 
   const handleUSDCManualTransferPurchase = useCallback(() => {
     setPurchaseMethod(PurchaseMethod.BALANCE)
     submitForm()
   }, [submitForm, setPurchaseMethod])
+
   return (
     <ModalForm className={cn(styles.modalRoot, { [styles.mobile]: isMobile })}>
       <ModalHeader
@@ -144,40 +139,34 @@ const RenderForm = ({
         onClose={onClose}
         showDismissButton={!isMobile}
       >
-        <ModalTitle icon={<IconCart />} title={messages.completePurchase} />
+        <ModalTitle
+          icon={<IconCart />}
+          title={
+            page === PurchaseContentPageType.GUEST_CHECKOUT
+              ? messages.guestCheckout
+              : messages.completePurchase
+          }
+        />
       </ModalHeader>
       <ModalContentPages
         contentClassName={styles.content}
         className={styles.content}
         currentPage={currentPageIndex}
       >
-        <GuestCheckout
-          metadata={metadata}
+        {!isSignedIn ? (
+          <GuestCheckoutPage
+            metadata={metadata}
+            price={price}
+            onClickSignIn={onClose}
+          />
+        ) : null}
+        <PurchaseContentPage
+          stage={stage}
+          isUnlocking={isUnlocking}
+          purchaseSummaryValues={purchaseSummaryValues}
           price={price}
-          onClickSignIn={onClose}
+          metadata={metadata}
         />
-        <>
-          <Flex p={isMobile ? 'l' : 'xl'} pb='m'>
-            <Flex direction='column' gap='xl' w='100%'>
-              <LockedContentDetailsTile
-                showLabel={false}
-                metadata={metadata}
-                owner={metadata.user}
-                disabled={isLinkDisabled}
-                earnAmount={USDC(price / 100)
-                  .round()
-                  .toShorthand()}
-              />
-              <PurchaseContentFormFields
-                stage={stage}
-                purchaseSummaryValues={purchaseSummaryValues}
-                isUnlocking={isUnlocking}
-                price={price}
-                metadata={metadata}
-              />
-            </Flex>
-          </Flex>
-        </>
         <USDCManualTransfer
           onClose={handleUSDCManualTransferClose}
           amountInCents={price}
@@ -185,7 +174,9 @@ const RenderForm = ({
         />
       </ModalContentPages>
       <ModalFooter className={styles.footer}>
-        {page === PurchaseContentPage.PURCHASE ? (
+        {page === PurchaseContentPageType.GUEST_CHECKOUT ? (
+          <GuestCheckoutFooter />
+        ) : page === PurchaseContentPageType.PURCHASE ? (
           <PurchaseContentFormFooter
             error={error}
             isUnlocking={isUnlocking}
@@ -266,22 +257,27 @@ export const PremiumContentPurchaseModal = () => {
         ? PurchaseVendor.COINFLOW
         : PurchaseVendor.STRIPE
     })
+
   const showGuestCheckout =
     guestCheckoutEnabled &&
     (!currentUser || (currentUser && !currentUser.handle)) &&
     initialValues.guestEmail === ''
+
   useEffect(() => {
     if (showGuestCheckout) {
-      dispatch(setPurchasePage({ page: PurchaseContentPage.GUEST_CHECKOUT }))
+      dispatch(
+        setPurchasePage({ page: PurchaseContentPageType.GUEST_CHECKOUT })
+      )
     }
   }, [showGuestCheckout, dispatch])
 
   // On form mount, pre-create user bank if needed and check for any usdc stuck
   // in root wallet from an aborted withdrawal
   useEffect(() => {
+    if (showGuestCheckout) return
     dispatch(eagerCreateUserBank())
     dispatch(startRecoveryIfNecessary())
-  }, [dispatch])
+  }, [dispatch, showGuestCheckout])
 
   const handleClose = useCallback(() => {
     // Don't allow closing if we're in the middle of a purchase
@@ -330,7 +326,7 @@ export const PremiumContentPurchaseModal = () => {
           validateOnChange={false}
           onSubmit={onSubmit}
         >
-          <RenderForm
+          <PremiumContentPurchaseForm
             contentId={contentId}
             metadata={metadata}
             onClose={handleClose}
