+++ packages/web/src/components/premium-content-purchase-modal/components/PurchaseContentFormFooter.tsx
@@ -220,8 +220,8 @@ export const PurchaseContentFormFooter = ({
     <>
       <Button
         disabled={isUnlocking}
-        color='lightGreen'
-        type={'submit'}
+        color={'lightGreen'}
+        type='submit'
         isLoading={isUnlocking}
         fullWidth
       >
